+++ src/models/conversation.ts
@@ -6,7 +6,7 @@ import { QueueItem } from './queue-item';
 import { EmergencyMessageType } from '../constants/message-types/emergency';
 import { NamespacedConversation } from '../namespaces/interfaces';
 import { Store } from './store';
-import { Message, Origin, MessageState } from './message';
+import { Message, Origin, MessageState, nextUniqueId } from './message';
 import { CALL_INFO, CONTENT_TYPE, REPLY_TO } from '../constants/headers';
 import { CALL_SUB, MULTIPART_MIXED, PIDF_LO, TEXT_PLAIN, TEXT_URI_LIST } from '../constants/content-types';
 import { Multipart, MultipartPart, CRLF } from './multipart';
@@ -33,7 +33,7 @@ export enum ConversationState {
 
 export interface SendMessageObject {
   /**
-   * Chat message to send
+   * Text message to be sent
    */
   text?: string,
   /**
@@ -242,7 +242,7 @@ export class Conversation {
             if (origin === Origin.REMOTE)
               this._setState(ConversationState.ERROR, origin)();
 
-            this._logger.error('Could not send SIP message.', evt);
+            this._logger.error('Could not send SIP message.', message, evt);
             reject(evt);
           },
         }
@@ -289,7 +289,6 @@ export class Conversation {
     if (this._heartbeatInterval || !this._isHeartbeatAllowed())
       return;
 
-    // TODO: Error handling if sendMessage fails
     this._heartbeatInterval = setInterval(
       this._store.getHeartbeatInterval(),
       () => this.sendMessage({
@@ -358,31 +357,35 @@ export class Conversation {
   /**
    * Starts the conversation
    * 
-   * @param text Text to be sent with the first SIP *MESSSAGE*\
-   * Defaults to "Start emergency call"
+   * This is basically a convenience function on top of `sendMessage` \
+   * It automatically sets the correct message type "START" \
+   * and defaults to text message "Start emergency call"
    */
-  start = (text?: string): Message => {
-    text = text ?? 'Start emergency call';
-
-    return this.sendMessage({
-      text,
+  start = (sendMessageObj?: SendMessageObject): Message => {
+    sendMessageObj = {
+      text: 'Start emergency call',
       type: EmergencyMessageType.START,
-    });
+      ...sendMessageObj,
+    }
+
+    return this.sendMessage(sendMessageObj);
   }
 
   /**
    * Ends the conversation
    * 
-   * @param text Text to be sent with the last SIP *MESSSAGE*\
-   * Defaults to "End emergency call"
+   * This is basically a convenience function on top of `sendMessage` \
+   * It automatically sets the correct message type "STOP" \
+   * and defaults to text message "Stop emergency call"
    */
-  stop = (text?: string): Message => {
-    text = text ?? 'End emergency call';
-
-    return this.sendMessage({
-      text,
+  stop = (sendMessageObj?: SendMessageObject): Message => {
+    sendMessageObj = {
+      text: 'Stop emergency call',
       type: EmergencyMessageType.STOP,
-    });
+      ...sendMessageObj,
+    }
+
+    return this.sendMessage(sendMessageObj);
   }
 
   /**
@@ -413,6 +416,7 @@ export class Conversation {
     const message: Message = {
       // if no message id is specified, use the internal sequence
       id: messageId ?? this._messageId++,
+      uniqueId: nextUniqueId(),
       origin: Origin.LOCAL,
       conversation: this,
       dateTime: new Date(),
@@ -578,6 +582,7 @@ export class Conversation {
 
       this._addNewMessage({
         id: this.mapper.getMessageIdFromHeaders(callInfoHeaders) as string,
+        uniqueId: nextUniqueId(),
         origin,
         conversation: this,
         dateTime: now,
@@ -616,6 +621,17 @@ export class Conversation {
    */
   addStateListener = (callback: (state: StateObject) => unknown) => {
     this._stateListeners.push(callback);
+  }
+
+  /**
+   * Unregisters a previously registered conversation state listener
+   * 
+   * @param callback Callback function that is called each time the conversation's state changes
+   */
+  removeStateListener = (callback: (state: StateObject) => unknown) => {
+    const index = this._stateListeners.indexOf(callback);
+    if (index !== -1)
+      this._stateListeners.splice(index, 1);
   }
 
   /**